feat: rates module for exchange rate APIs
