+++ crates/interledger-api/Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/interledger-rs/interledger-rs"
 interledger-packet = { path = "../interledger-packet", version = "^0.4.0", default-features = false }
 interledger-http = { path = "../interledger-http", version = "^0.4.0", default-features = false }
 interledger-ildcp = { path = "../interledger-ildcp", version = "^0.4.0", default-features = false }
+interledger-rates = { path = "../interledger-rates", version = "^0.4.0", default-features = false }
 interledger-router = { path = "../interledger-router", version = "^0.4.0", default-features = false }
 interledger-service = { path = "../interledger-service", version = "^0.4.0", default-features = false }
 interledger-service-util = { path = "../interledger-service-util", version = "^0.4.0", default-features = false }
