+++ crates/interledger-api/src/routes/accounts.rs
@@ -7,12 +7,13 @@ use interledger_errors::*;
 use interledger_http::{deserialize_json, HttpAccount, HttpStore};
 use interledger_ildcp::IldcpRequest;
 use interledger_ildcp::IldcpResponse;
+use interledger_rates::ExchangeRateStore;
 use interledger_router::RouterStore;
 use interledger_service::{
     Account, AccountStore, AddressStore, IncomingService, OutgoingRequest, OutgoingService,
     Username,
 };
-use interledger_service_util::{BalanceStore, ExchangeRateStore};
+use interledger_service_util::BalanceStore;
 use interledger_settlement::core::{types::SettlementAccount, SettlementClient};
 use interledger_spsp::{pay, SpspResponder};
 use interledger_stream::{PaymentNotification, StreamNotificationsStore};
