+++ crates/interledger-rates/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "interledger-rates"
+version = "0.4.0"
+authors = ["Kincaid O'Neil <kincaidoneil@users.noreply.github.com>"]
+edition = "2018"
+
+# TODO Add other metadata here!
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = "0.1.22"
+futures = { version = "0.3.1", default-features = false }
+interledger-errors = { path = "../interledger-errors", version = "^0.1.0" }
+log = { version = "0.4.8", default-features = false }
+once_cell = "1.3.1"
+reqwest = { version = "0.10.0", default-features = false, features = ["default-tls", "json"] }
+secrecy = { version = "0.6", default-features = false, features = ["alloc", "serde"] }
+serde = { version = "1.0.101", default-features = false, features = ["derive"]}
+tokio = { version = "0.2.6", default-features = false, features = ["macros", "time"] }
