+++ crates/interledger-rates/README.md
@@ -1,0 +1,3 @@
+# interledger-rates
+
+Utilities for fetching and caching exchange rates from external APIs, which supports CoinCap and CryptoCompare rate backends.
