--- crates/interledger-service-util/src/exchange_rate_providers/mod.rs
@@ -1,7 +1,0 @@
-/// Exchange rate provider for [CoinCap](https://coincap.io/)
-mod coincap;
-/// Exchange rate provider for [CryptoCompare](https://www.cryptocompare.com/). REQUIRES [API KEY](https://min-api.cryptocompare.com/).
-mod cryptocompare;
-
-pub use coincap::*;
-pub use cryptocompare::*;
