+++ crates/interledger-store/Cargo.toml
@@ -26,6 +26,7 @@ interledger-packet = { path = "../interledger-packet", version = "^0.4.0", defau
 interledger-btp = { path = "../interledger-btp", version = "^0.4.0", default-features = false }
 interledger-ccp = { path = "../interledger-ccp", version = "^0.3.0", default-features = false }
 interledger-http = { path = "../interledger-http", version = "^0.4.0", default-features = false }
+interledger-rates = { path = "../interledger-rates", version = "^0.4.0", default-features = false }
 interledger-router = { path = "../interledger-router", version = "^0.4.0", default-features = false }
 interledger-service = { path = "../interledger-service", version = "^0.4.0", default-features = false }
 interledger-service-util = { path = "../interledger-service-util", version = "^0.4.0", default-features = false }
