+++ crates/interledger/Cargo.toml
@@ -15,6 +15,7 @@ node = [
     "ccp",
     "http",
     "ildcp",
+    "rates",
     "router",
     "service-util",
     "settlement",
@@ -28,6 +29,7 @@ btp = ["interledger-btp"]
 ccp = ["interledger-ccp"]
 http = ["interledger-http"]
 ildcp = ["interledger-ildcp"]
+rates = ["interledger-rates"]
 router = ["interledger-router"]
 service-util = ["interledger-service-util"]
 settlement = ["interledger-settlement" ]
@@ -45,6 +47,7 @@ interledger-http = { path = "../interledger-http", version = "^0.4.0", optional 
 interledger-ildcp = { path = "../interledger-ildcp", version = "^0.4.0", optional = true, default-features = false }
 interledger-packet = { path = "../interledger-packet", version = "^0.4.0", default-features = false }
 interledger-errors = { path = "../interledger-errors", version = "^0.1.0", default-features = false }
+interledger-rates = { path = "../interledger-rates", version = "^0.4.0", optional = true, default-features = false }
 interledger-router = { path = "../interledger-router", version = "^0.4.0", optional = true, default-features = false }
 interledger-service = { path = "../interledger-service", version = "^0.4.0", default-features = false }
 interledger-service-util = { path = "../interledger-service-util", version = "^0.4.0", optional = true, default-features = false }
